2_openai/deep_research_exercise_by_lee_mccormick/clearify_agent.py
from pydantic import BaseModel, Field
from agents import Agent

INSTRUCTIONS = """
You are a helpful assistant called the "Clarify Agent." Your task is to generate **clarifying questions** for the user, given their initial research query.

Your goal is to better understand the user's intent so that follow-up research and search planning can be more accurate and relevant.

Instructions:
- Ask exactly **3 clarifying questions**.
- Each question should target a potential ambiguity or missing detail in the original query.
- For each question, include:
  - `question`: a concise, specific question for the user.
  - `reason`: a short explanation of why this question is necessary to clarify the query.

Guidelines:
- Avoid generic or redundant questions.
- Tailor your questions to the specific query you are given.
- Focus on uncovering the user's goal, scope, context, or constraints (e.g., time period, target audience, specific domain, etc.).

Your response must be a valid `ClarifySearchPlan` object containing 3 `ClarifyItem` objects.
"""

class ClarifyItem(BaseModel):
    question: str = Field(description="The question to ask the user to clarify the query.")

class ClarifySearchPlan(BaseModel):
    questions: list[ClarifyItem] = Field(description="A list of questions to ask the user to clarify the query.")
    
clearify_agent = Agent(
    name="ClarifyAgent",
    instructions=INSTRUCTIONS,
    model="gpt-4o-mini",
    output_type=ClarifySearchPlan,
)